+++ src/module/RelationData.ts
@@ -1,4 +1,16 @@
-import BaseData, { bindOption, bindType, loadFunctionType } from "../data/BaseData"
+import { upperCaseFirstChar } from "complex-utils"
+import BaseData, { loadFunctionType } from "../data/BaseData"
+
+export type bindLifeType = 'load' | 'update'
+
+export type unbindType = (life?: string[]) => void
+
+export type bindType = (depend: BaseData, self: BaseData, success: boolean, life: bindLifeType, unbind: unbindType) => void
+
+export interface bindOption {
+  life?: bindLifeType
+  active?: boolean
+}
 
 export type nextTypeFunction<D> = (status: 'success' | 'fail', depend: D, res: unknown) => unknown
 
@@ -36,6 +48,65 @@ export interface RelationDataInitOption {
 class RelationData {
   static $name = 'RelationData'
   static $formatConfig = { name: 'Data:RelationData', level: 10, recommend: false }
+  static $bindDependByActive(self: BaseData, depend: BaseData, bind: bindType, from: string, success: boolean, life: bindLifeType, unbind: () => void, active?: boolean) {
+    let sync = true
+    if (active && !self.$isActive()) {
+      // 需要判断激活状态且当前状态为未激活时不同步触发
+      sync = false
+    }
+    if (sync) {
+      bind(depend, self, success, life, unbind)
+    } else {
+      // 设置主数据被激活时触发bind函数
+      // 设置相同id,使用replace模式，需要注意的是当函数变化后开始的函数可能还未被触发
+      self.$onLife('actived', {
+        id: depend.$getId('BindLife' + upperCaseFirstChar(from)),
+        once: true,
+        replace: true,
+        data: () => {
+          bind(depend, self, success, life, unbind)
+        }
+      })
+    }
+  }
+  static $bindDependByLife(self: BaseData, depend: BaseData, bind: bindType, life: bindLifeType, lifeDict: Record<string, PropertyKey> = {}, {
+    active, // 是否只在激活状态下触发
+  }: bindOption = {}) {
+    if (active === undefined && self.$active.auto) {
+      // 自动激活模式下，默认进行激活的判断
+      active = true
+    }
+    const failLifeName = life === 'load' ? 'loadFail' : 'updateFail'
+    const successLifeName = life === 'load' ? 'loaded' : 'updated'
+    const currentStatus = depend.$getStatus(life)
+    const unbind: unbindType = function(lifeList?: string[]) {
+      for (const lifeName in lifeDict) {
+        if (lifeList === undefined || lifeList.indexOf(lifeName) > -1) {
+          depend.$offLife(lifeName, lifeDict[lifeName])
+        }
+      }
+    }
+    lifeDict[successLifeName] = depend.$onLife(successLifeName, {
+      data: () => {
+        this.$bindDependByActive(self, depend, bind, successLifeName, true, life, unbind, active)
+      }
+    }) as PropertyKey
+    lifeDict[failLifeName] = depend.$onLife(failLifeName, {
+      data: () => {
+        this.$bindDependByActive(self, depend, bind, failLifeName, false, life, unbind, active)
+      }
+    }) as PropertyKey
+    if (currentStatus === 'success') {
+      this.$bindDependByActive(self, depend, bind, successLifeName, true, life, unbind, active)
+    } else if (currentStatus === 'fail') {
+      this.$bindDependByActive(self, depend, bind, failLifeName, false, life, unbind, active)
+    }
+  }
+  static $bindDepend(self: BaseData, depend: BaseData, bind: bindType, option: bindOption = {}) {
+    const lifeDict: Record<string, PropertyKey> = {}
+    this.$bindDependByLife(self, depend, bind, 'load', lifeDict, option)
+    this.$bindDependByLife(self, depend, bind, 'update', lifeDict, option)
+  }
   depend?: {
     type: 'sync' | 'order'
     list: dependValueType[]
@@ -63,7 +134,7 @@ class RelationData {
         item.args = [false]
       }
       if (item.bind) {
-        self.$bindDepend(item.data, item.bind.data, item.bind)
+        RelationData.$bindDepend(self, item.data, item.bind.data, item.bind)
       }
       return item as dependValueType
     }